[UPDATE] 優化業務流程，精簡冗餘業務，更新SQl文件

+++ src/main/java/com/moonlit/generator/generator/entity/vo/TableFieldVO.java
@@ -39,7 +39,7 @@ public class TableFieldVO implements Serializable {
     /**
      * 順序
      */
-    private Long ordinalPosition;
+    private Integer ordinalPosition;
 
     /**
      * 備注
